move attribute to separate class file (fxcop)
