+++ src/Microsoft.ApplicationInsights/Extensibility/Implementation/Endpoints/EndpointName.cs
@@ -1,8 +1,5 @@
 ﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Endpoints
 {
-    using System;
-    using System.Reflection;
-
     /// <summary>
     /// These enums represent all possible endpoints within application insights infrastructure.
     /// </summary>
@@ -23,33 +20,5 @@
         /// <summary>Snapshot debugger service.</summary>
         [EndpointMeta(ExplicitName = "SnapshotEndpoint", EndpointPrefix = Constants.SnapshotPrefix, Default = Constants.SnapshotEndpoint)]
         Snapshot,
-    }
-
-    /// <summary>
-    /// Defines meta data for possible endpoints.
-    /// </summary>
-    internal class EndpointMetaAttribute : Attribute
-    {
-        /// <summary>Gets or sets the explicit name for overriding an endpoint within a connection string.</summary>
-        public string ExplicitName { get; set; }
-
-        /// <summary>Gets or sets the prefix (aka subdomain) for an endpoint.</summary>
-        public string EndpointPrefix { get; set; }
-
-        /// <summary>Gets or sets the default classic endpoint.</summary>
-        public string Default { get; set; }
-
-        public static EndpointMetaAttribute GetAttribute(EndpointName enumValue)
-        {
-#if NETSTANDARD1_3
-            Type type = enumValue.GetType();
-            string name = Enum.GetName(type, enumValue);
-            return type.GetRuntimeField(name).GetCustomAttribute<EndpointMetaAttribute>();
-#else
-            Type type = enumValue.GetType();
-            string name = Enum.GetName(type, enumValue);
-            return type.GetField(name).GetCustomAttribute<EndpointMetaAttribute>();
-#endif
-        }
     }
 }
